g7

+++ pybricks/pyBricks1/main.py
@@ -194,11 +194,11 @@ async def get_the_vegetables_at_the_market():
 
     # position wall near compose area
     await WheelController.wheel_slight_left_turn()
-    await WheelController.move_wheels_backward_in_straight_line(float(230))
+    await WheelController.move_wheels_backward_in_straight_line(float(190))
     await WheelController.wheel_slight_right_turn()
     await WheelController.move_wheels_forward_in_straight_line(float(150))
     await WheelController.wheel_u_turn_right()
-    await multitask(WheelController.move_wheels_backward_in_straight_line(float(255), with_brake=True),
+    await multitask(WheelController.move_wheels_backward_in_straight_line(float(145), with_brake=True),
                     GripperController.grip_element_using_both_arms())
